+++ spring-petclinic-api-gateway/src/main/java/org/springframework/samples/petclinic/api/application/VisitsServiceClient.java
@@ -16,6 +16,7 @@
 package org.springframework.samples.petclinic.api.application;
 
 import org.springframework.samples.petclinic.api.dto.Visits;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 import org.springframework.web.reactive.function.client.WebClient;
 import reactor.core.publisher.Mono;
@@ -30,28 +31,24 @@ import static java.util.stream.Collectors.joining;
 @Component
 public class VisitsServiceClient {
 
-    // Could be changed for testing purpose
-    private String hostname = "http://visits-service/";
-
     private final WebClient.Builder webClientBuilder;
+    private final String visitsUrl;
 
-    public VisitsServiceClient(WebClient.Builder webClientBuilder) {
+    public VisitsServiceClient(WebClient.Builder webClientBuilder,
+                               @Value("${petclinic.services.visits-url}") String visitsUrl) {
         this.webClientBuilder = webClientBuilder;
+        this.visitsUrl = visitsUrl;
     }
 
     public Mono<Visits> getVisitsForPets(final List<Integer> petIds) {
         return webClientBuilder.build()
             .get()
-            .uri(hostname + "pets/visits?petId={petId}", joinIds(petIds))
+            .uri(visitsUrl + "/pets/visits?petId={petId}", joinIds(petIds))
             .retrieve()
             .bodyToMono(Visits.class);
     }
 
     private String joinIds(List<Integer> petIds) {
         return petIds.stream().map(Object::toString).collect(joining(","));
-    }
-
-    void setHostname(String hostname) {
-        this.hostname = hostname;
     }
 }