+++ spring-petclinic-api-gateway/src/test/java/org/springframework/samples/petclinic/api/application/VisitsServiceClientIntegrationTest.java
@@ -26,8 +26,8 @@ class VisitsServiceClientIntegrationTest {
     @BeforeEach
     void setUp() {
         server = new MockWebServer();
-        visitsServiceClient = new VisitsServiceClient(WebClient.builder());
-        visitsServiceClient.setHostname(server.url("/").toString());
+        visitsServiceClient = new VisitsServiceClient(WebClient.builder(),
+            server.url("/").toString().replaceAll("/$", ""));
     }
 
     @AfterEach