+++ spring-petclinic-customers-service/src/main/java/org/springframework/samples/petclinic/customers/config/RestTemplateConfig.java
@@ -1,7 +1,6 @@
 package org.springframework.samples.petclinic.customers.config;
 
 import org.springframework.boot.restclient.RestTemplateBuilder;
-import org.springframework.cloud.client.loadbalancer.LoadBalanced;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.client.RestTemplate;
@@ -12,7 +11,6 @@ import java.time.Duration;
 public class RestTemplateConfig {
 
     @Bean
-    @LoadBalanced
     RestTemplate restTemplate(RestTemplateBuilder builder) {
         return builder
             .connectTimeout(Duration.ofSeconds(2))